Check bitfield if peer has desired piece
The change to the enqueueing to put a work item to the front of the
queue rather than the back of the queue was necessary to avoid the
following situation:
- theer are two tasks for processing work items on the shared queue
- there are two work items left on the shared queue
- each task has only one work item left on the queue that they're able
  to process
- the ordering of work items on the queue is such that:
    - one task gets the work item from the front of the queue that it
      can't process (but the other task could), puts it onto the back of
      the queue, and yields
    - the other task gets the work item from the front of the queue that
      it can't process (but the other task could), puts it onto the back
      of the queue, and yields
    - this repeats indefinitely, and neither of the two work items on
      the queue ever get processed

As a consequence of the above change, the async sleep is required to
make sure that other tasks have a chance of being scheduled after the
work item has been put back onto the shared queue. Ie, so then the same
task doesn't pick up the same work item that it couldn't process.