Remove use of `BufReader` to read response handshake
After idly reading the [docs for
`BufReader`](https://docs.rs/tokio/latest/tokio/io/struct.BufReader.html)
I came across this line:
```
When the BufReader is dropped, the contents of its buffer will be
discarded. Creating multiple instances of a BufReader on the same stream
can cause data loss.
```

This rang a warning bell in my mind, as there's multiple places in the
code which is creating a `BufReader` out of the same `TcpStream`:
- `Client::handshake()`
- `Message::deserialise()`

Noting the race condition mentioned in #3, and not being satisfied with
the sleep added in the test in d6c4a80 (where, without the sleep,
hanging seemingly non-deterministically occurred in
`Message::deserialise()`), on a whim, I tried removing one of the
`BufReader` instances that were involved in that test to see how that
fared without the sleep.

This turned out to actually work, where the test no longer needs the 5ms
sleep to consistently pass.